refactor(porygon-typeorm): update modules folder
