+++ packages/porygon-typeorm/src/persistence/builders/domain/buildPorygonTypeOrmEntityModuleSymbolsMap.spec.ts
@@ -1,12 +1,12 @@
 import { afterAll, beforeAll, describe, expect, it, jest } from '@jest/globals';
 
-jest.mock('../../foundation/calculations/hashString');
+jest.mock('../../../foundation/calculations/hashString');
 
 import { Newable } from '@cuaklabs/iocuak';
 
-import { hashString } from '../../foundation/calculations/hashString';
-import { PorygonTypeOrmEntityModuleSymbolsMap } from '../models/domain/PorygonTypeOrmEntityModuleSymbolsMap';
-import { PorygonTypeOrmSourceModuleSymbolsMap } from '../models/domain/PorygonTypeOrmSourceModuleSymbolsMap';
+import { hashString } from '../../../foundation/calculations/hashString';
+import { PorygonTypeOrmEntityModuleSymbolsMap } from '../../models/domain/PorygonTypeOrmEntityModuleSymbolsMap';
+import { PorygonTypeOrmSourceModuleSymbolsMap } from '../../models/domain/PorygonTypeOrmSourceModuleSymbolsMap';
 import { buildPorygonTypeOrmEntityModuleSymbolsMap } from './buildPorygonTypeOrmEntityModuleSymbolsMap';
 
 describe(buildPorygonTypeOrmEntityModuleSymbolsMap.name, () => {
